NCBI merge databases operation.
The pipeline splign+compart+bedtools decribed by Jorge Vieira is now
supported. Several important changes were done to support this workflow:
	- EntityValidator now supports specific entity validation deactivation.
	- The validate method of EntityValidator now does not returns any
value.
	- RepositoryManager now has a method for validating entities.
	- AbstractBinariesExecutor now supports the use of callback to manage
the lines readed from the default and error streams of the executed
commands. This is useful to reduce the memory consumption with large
outputs.
	- While an operation is executed, an "operation running" dialog is
showed. It would be interesting to add support for cancelling
operations.