Bug 133. Not all sequences are retrieved. Repository tree improvement.
This bug was causing that BLAST operations using a multiple sequence
FASTA file as query only generate results for the first sequence.

While solving this bug, an bug in the repository tree view was detected.
This bug hanged the application when several files were deleted (as
occurs in the BLAST operations). Repository tree was improved and, in
addition to the bugfix, now the items are sorted by alphabetical order.