Fasta sequence renaming improvement
New options added to the fasta reformat operation. Now it is possible to
keep the sequence name and to include or not a counter after the prefix
when using the "Prefix" sequence name reformat method. In addition, the
"Separator" parameter allows user to select the separator to use when
several parts (e.g. prefix, counter, etc.) are used to create the new
name.

Minor source formating.